% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_backup_cmd.r, R/Data_get_workflow.r
\name{Data_backup_cmd}
\alias{Data_backup_cmd}
\alias{Data_get_workflow}
\title{backup do for Master files}
\usage{
Data_backup_cmd(wd)

Data_get_workflow()
}
\description{
not longer used

helper to work with bulk download processing  workflow
}
\examples{
## Not run:

## End(**Not run**)
## Not run:
## End(**Not run**)
}
\author{
ILO / bescond

ILO / bescond
}
\keyword{ILO,}
\keyword{microdataset,}
\keyword{preprocessing}
\keyword{processing}
